refactor: extract login to its own module
